RPC: Get skeleton to compile
I made some wrong assumptions before. I think for the first draft I'm
just going to dump a bunch of code in stereo_tri_rpc and then refactor
latter. I can't seem to gather the whole scope of RPC into my brain at
one time to actually make a good effort on the API.

In fact, it's a bit of a bastardization by me to put the RPC camera
model as part of VW's camera model. It's because I can't implement
most of the functions in that API. Sessions is also going to have to
be rewritten so that we don't return shared_ptr pointers. I need to
get down to a point where the triangulation can freely make copies of
the camera model and not interface through a mutex'd smart pointer.